pyproject.toml
[project]
name = "x-streamlit-sanbox"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2.11",
    "langchain-openai>=1.1.11",
    "langgraph>=1.1.0",
    "python-dotenv>=1.2.2",
    "streamlit>=1.55.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.15.5",
]
[tool.ruff.format]
quote-style = "single"
indent-style = "space"
[tool.ruff.lint]
select = ["I"]
ignore = []
